+++ src/MapCache.hpp
@@ -5,5 +5,6 @@ namespace versus {
 bool mapCached(int64_t levelID);
 void ensureMapCached(int64_t levelID, Done callback);
 void preloadBattleAssets();
+void preloadRunnerAssets();
 GJGameLevel* cachedLevel(int64_t levelID);
 }